Ensure API safety

+++ src/shift/shift.controller.ts
@@ -12,18 +12,31 @@ import { DocumentReference } from 'firebase-admin/lib/firestore';
 import { firestore } from 'firebase-admin';
 import FieldValue = firestore.FieldValue;
 
+function customIsValidDate(date: string) {
+  const regex = '^\\d{4}-\\d{2}-\\d{2}$';
+  if (!date.match(regex)) return false;
+  const d = new Date(date);
+  const dNum = d.getTime();
+  if (!dNum && dNum !== 0) return false;
+  return d.toISOString().slice(0, 10) === date;
+}
+
 @Controller('shift')
 export class ShiftController {
   @Delete('delete/:id')
   async deleteShift(@Param('id') id: string, @Query('admin') admin: string) {
     const shiftData = await cloud_firestore.doc(`shifts/${id}`).get();
+    if (!shiftData.exists) return 'Shift does not exist.';
+    const employee = await (
+      shiftData.data()['employee'] as DocumentReference
+    ).get();
+    if (!employee.exists) return 'Employee of shift does not exist';
+    const adminData = await cloud_firestore.doc(`users/${admin}`).get();
+    if (!adminData.exists) return 'Admin does not exist';
     const admins = await cloud_firestore
       .collection('users')
       .where('isAdmin', '==', true)
       .get();
-    const employee = await (
-      shiftData.data()['employee'] as DocumentReference
-    ).get();
     await cloud_firestore.collection('changes').add({
       type: 'DELETE',
       employee: employee.ref,
@@ -99,6 +112,13 @@ export class ShiftController {
     @Query('admin') admin: string,
     @Body() shiftDetails: ShiftDetails,
   ) {
+    const employee = await cloud_firestore
+      .doc(`users/${shiftDetails.employee}`)
+      .get();
+    if (!employee.exists) return 'Employee does not exist';
+    const adminData = await cloud_firestore.doc(`users/${admin}`).get();
+    if (!adminData.exists) return 'Admin does not exist';
+    if (!customIsValidDate(shiftDetails.date)) return 'Invalid date format';
     await cloud_firestore.collection('shifts').add({
       employee: cloud_firestore.doc(`users/${shiftDetails.employee}`),
       time: shiftDetails.time,
@@ -112,9 +132,6 @@ export class ShiftController {
       employee: cloud_firestore.doc(`users/${shiftDetails.employee}`),
       timestamp: FieldValue.serverTimestamp(),
     });
-    const employee = await cloud_firestore
-      .doc(`users/${shiftDetails.employee}`)
-      .get();
     const admins = await cloud_firestore
       .collection('users')
       .where('isAdmin', '==', true)
@@ -180,7 +197,15 @@ export class ShiftController {
     @Query('admin') admin: string,
     @Body() shiftDetails: ShiftDetails,
   ) {
+    const employee = await cloud_firestore
+      .doc(`users/${shiftDetails.employee}`)
+      .get();
+    if (!employee.exists) return 'Employee does not exist';
+    const adminData = await cloud_firestore.doc(`users/${admin}`).get();
+    if (!adminData.exists) return 'Admin does not exist';
+    if (!customIsValidDate(shiftDetails.date)) return 'Invalid date format';
     const oldShift = await cloud_firestore.doc(`shifts/${id}`).get();
+    if (!oldShift.exists) return 'Shift does not exist';
     await cloud_firestore.doc(`shifts/${id}`).set({
       employee: cloud_firestore.doc(`users/${shiftDetails.employee}`),
       time: shiftDetails.time,
@@ -196,9 +221,6 @@ export class ShiftController {
       employee: cloud_firestore.doc(`users/${shiftDetails.employee}`),
       timestamp: FieldValue.serverTimestamp(),
     });
-    const employee = await cloud_firestore
-      .doc(`users/${shiftDetails.employee}`)
-      .get();
     const admins = await cloud_firestore
       .collection('users')
       .where('isAdmin', '==', true)
@@ -279,6 +301,6 @@ export class ShiftController {
 
 interface ShiftDetails {
   employee: string;
-  time: string;
+  time: 'morning' | 'evening';
   date: string;
 }
